feat: create base JSON library to reduce reliance on json-api library in widgets library
Create a base JSON library that contains the most basic code required for interacting
with a JSON API and use that in the widgets library instead of the json-api library. This
should allow users to utilize whatever JSON format that they want, provided they
follow the basic guidelines laid out in the base-json library.