fix: generating routes and sidenav items
Narrow includes conditions to include slashes ("/") in routes around expected module
names and types, and include both names and types (in places where needed). This
fixes issue with overlapping sidenavs and routes for modules with similar names. For
example, the json library was including base-json sidenav items since "json" is
included in "base-json". Narrowing conditions seems to resolve this.