fix: issue updating sidenavItems
Replace routerSub with setter for sidenavItems and re-initialize sidenavItems there.
This fixes issue with sidenav items not updating properly in the component when the
route changes (or at least getting initialized). Biggest problem was when navigating
from json libraries to widgets library. While the sidenavItems appeared correctly it is
likely that the re-initialization of sidenavItems occurred before the sidenavItems
changed, resulting in issues opening items beyond the 3rd one (the same number of
items in the json libraries).