fix: updating models
Replace Reflect.getMetadata('Attribute', model) statement with
model[AttributeMetadataIndex] to fix issue gettting attribute metadata in
updateRecord method. I don't know where I came up with the original code for this,
but the attribute metadata was not correct when using this code and as a result,
the body of the request was not including the correct attribute data. The updated code follows the same structure as the save method for models and this seems to fix
the issue with the attribute data.