refactor: lots of changes to simplify code
Update code to use more inheritance so we can remove a lot of duplicate code. Also
rename components to (hopefully) make it somewhat easier to follow the chain of inheritance. Need to figure out what impacts if any the heavy use of inheritance has
on debugging issues or causing unforeseen side effects.