fix: paginator theme in toolbar
Paginator background and font colors do not match when included in toolbar, causing the toolbar to look like it is split in multiple parts. Add a new mixin to set styles to match the rest of the toolbar, and update theme switcher to call mixin. Also replace css styles with scss for component.