fix: selection when data changes
If one or more rows selected when data changes, the selection on the component remains until user selects one or more rows. This causes any buttons in toolbar to remain enabled when data changes (even if they should not be). Add logic to clear selection and emit boolean to disable any buttons that should be disabled.