fix(documentation): layout
If the main content area does not fill the screen, then the footer is rendered well above the bottom of the page and does not "stick" to the bottom in those instances. Update styles to ensure footer "sticks" to the bottom of the page even when the content does not fill the space.