chore(widgets): refactor filter drop down
Replace MatMenu with MatDialog positioned below search field and rendered with custom background to make dialog look like a menu. This mainly fixes an issue with font theme in drop down (which does not change when switching to dark theme; leaving dark text on dark background). This is due to the fact that the content is not rendered inside of a MatMenuItem, which applies appropriate styles to text when theme is switched. Forcing the content into a MatMenuItem doesn't really make sense since it is a form. So a dialog is used since it can be made to look like a menu (and the look of a menu is really all I'm going for here). This is internal functionality only, and does not affect how component is used.