fix(widgets): closing dialog with keyboard
It seems that multiple dialogs are opened and stacked on top of each other resulting in an issue when closing the dialog with the escape key causing an empty dialog to appear under dialog with form details. Additional logic added to ensure only a single dialog is opened, which seems to fix issue closing dialog with escape key.